Fixes ordering of necessary backup.
It now depends on the order of the keys in the configuration file.
Beforehand, min() was called on the list of tuples of necessary backups,
the tuples being (name, interval). As the ordering of lists depends only
on the first element, it was merely an alphabetical ordering of the
interval names. As "daily" < "hourly" < "minutely" (alphabetical
ordering, but len(str) would do the same), this (un)luckily worked with
the default configuration file.