Fix ordering of interval tags.
Two things were wrong. First, the interval tags were stored in a
dictionary, which of course has no consistent ordering. To avoid this
issue, the dict() was replaced by collections.OrderedDict(). This is
also used for the "sections" dict, so we have consistent ordering
information about all keys in a section, too.

The second issue was in the algorithm for searching the right backup
interval. When a match was found, a "break" followed. But this break of
course only broke out of the inner loop, the outer loop continued
searching, so in the end, the result was something obscure. This was
fixed with a "exitloop" variable that terminates the outer loop after
the break in the inner one.

I got NO idea how the code even remotely worked before this.