Add support for multiple "=" per line.
Before, when multiple "=" were present, unpacking after split() failed
and an error was raised. Now, partition() is used instead of split(),
which splits at the first occurence of "=" and returns a tuple
containing the first part, the separator ("=") and the second part. By
unpacking this tuple and just ignoring the separator we have the exact
same functionality when only one "=" is present, but e.g.

    device="UUID=abc"

will be parsed as

    (device, "UUID=abc")

which is desired behaviour.