Split __init__.py into separate files.
The logging functionality stays in __init__.py

Command line arguments are handled by daemon.py, which is the file that
is called by the rbackupd script.

The parsing of the configuration file is done in backupmanager.py, which
also initializes all tasks and starts all relevant processes for
monitoring backups.