Fix bug with removing already removed backups.
This happended when a backup was expired because it was too old AND
because of the amount of backups of that interval. As the expired
backups regarding each criterium were simply appended on each other, a
backup from both critera would appear twice in the list an therefore the
program would try to remove it twice, failing the second time.